+++ src/main/java/chess/database/GameStateGenerator.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.database;
 
 import java.util.List;
 import java.util.Map;
