+++ src/main/java/chess/database/dao/JdbcBoardDao.java
@@ -1,12 +1,12 @@
-package chess.dao;
+package chess.database.dao;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import chess.database.BoardDto;
-import chess.database.PieceDto;
-import chess.database.PointDto;
-import chess.database.RouteDto;
+import chess.database.dto.BoardDto;
+import chess.database.dto.PieceDto;
+import chess.database.dto.PointDto;
+import chess.database.dto.RouteDto;
 
 public class JdbcBoardDao implements BoardDao {
