+++ src/main/java/chess/database/dto/BoardDto.java
@@ -1,4 +1,4 @@
-package chess.database;
+package chess.database.dto;
 
 import java.util.Map;
 import java.util.stream.Collectors;
