+++ src/main/java/chess/database/dto/GameStateDto.java
@@ -1,4 +1,4 @@
-package chess.database;
+package chess.database.dto;
 
 import chess.domain.game.GameState;
