+++ src/main/java/chess/database/dto/PointDto.java
@@ -1,4 +1,4 @@
-package chess.database;
+package chess.database.dto;
 
 import java.util.Objects;
