+++ src/main/java/chess/domain/board/Board.java
@@ -7,9 +7,6 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Stream;
 
-import chess.database.BoardDto;
-import chess.database.PieceDto;
-import chess.database.PointDto;
 import chess.domain.Color;
 import chess.domain.piece.Empty;
 import chess.domain.piece.Piece;
