+++ src/test/java/chess/domain/game/FinishedTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.dto.Arguments;
 import chess.domain.Color;
 import chess.domain.board.BoardFixtures;
+import chess.dto.Arguments;
 
 class FinishedTest {
