+++ src/test/java/chess/domain/game/RunningTest.java
@@ -7,13 +7,13 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.dto.Arguments;
 import chess.domain.Color;
 import chess.domain.board.BoardFixtures;
 import chess.domain.board.Point;
 import chess.domain.piece.King;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
+import chess.dto.Arguments;
 
 class RunningTest {
